smoke-tests/requirements.txt
# Core OTel SDK + OTLP exporter — required by all smoke tests
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-http>=1.22.0

# Env file loading
python-dotenv>=1.0.0

# ── E2E auto-instrumentation tests (82, 83, 84, 85) ──────────────────────────
flask>=3.0.0
sqlalchemy>=2.0.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
django>=5.0.0
httpx>=0.27.0
aiohttp>=3.9.0

opentelemetry-instrumentation-flask>=0.43b0
opentelemetry-instrumentation-sqlalchemy>=0.43b0
opentelemetry-instrumentation-requests>=0.43b0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-django>=0.43b0
opentelemetry-instrumentation-httpx>=0.43b0
opentelemetry-instrumentation-aiohttp-client>=0.43b0
opentelemetry-instrumentation-urllib3>=0.43b0
